+++ src/utilities/api/queries.js
@@ -11,7 +11,7 @@ export const GET_COMMAND_PARAMETERS_SCHEMAS_QUERY = 'GET_COMMAND_PARAMETERS_SCHE
 export const GET_ACCOUNT_TOKENS_QUERY = 'GET_ACCOUNT_TOKENS_QUERY';
 export const GET_ACCOUNT_TOKENS_FULL_DATA_QUERY = 'GET_ACCOUNT_TOKENS_FULL_DATA_QUERY';
 export const GET_ACCOUNT_TRANSACTIONS_QUERY = 'GET_ACCOUNT_TRANSACTIONS_QUERY';
-export const GET_ACCOUNT_POOL_TRANSACTIONS_QUERY = 'GET_ACCOUNT_POOL_TRANSACTIONS_QUERY';
+export const GET_TRANSACTION_POOL_QUERY = 'GET_TRANSACTION_POOL_QUERY';
 export const GET_TOKENS_METADATA_QUERY = 'GET_TOKENS_METADATA_QUERY';
 export const GET_SUPPORTED_TOKENS_QUERY = 'GET_SUPPORTED_TOKENS_QUERY';
 export const INVOKE_QUERY = 'INVOKE_QUERY';
